Add SMS & Windwalker standards

+++ php/squizlabs/php_codesniffer/CodeSniffer/Reporting.php
@@ -8,8 +8,8 @@
  * @package   PHP_CodeSniffer
  * @author    Gabriele Santini <gsantini@sqli.com>
  * @author    Greg Sherwood <gsherwood@squiz.net>
- * @copyright 2009 SQLI <www.sqli.com>
- * @copyright 2006-2012 Squiz Pty Ltd (ABN 77 084 670 600)
+ * @copyright 2009-2014 SQLI <www.sqli.com>
+ * @copyright 2006-2014 Squiz Pty Ltd (ABN 77 084 670 600)
  * @license   https://github.com/squizlabs/PHP_CodeSniffer/blob/master/licence.txt BSD Licence
  * @link      http://pear.php.net/package/PHP_CodeSniffer
  */
@@ -27,8 +27,8 @@ if (is_file(dirname(__FILE__).'/../CodeSniffer.php') === true) {
  * @package   PHP_CodeSniffer
  * @author    Gabriele Santini <gsantini@sqli.com>
  * @author    Greg Sherwood <gsherwood@squiz.net>
- * @copyright 2009 SQLI <www.sqli.com>
- * @copyright 2006-2012 Squiz Pty Ltd (ABN 77 084 670 600)
+ * @copyright 2009-2014 SQLI <www.sqli.com>
+ * @copyright 2006-2014 Squiz Pty Ltd (ABN 77 084 670 600)
  * @license   https://github.com/squizlabs/PHP_CodeSniffer/blob/master/licence.txt BSD Licence
  * @version   Release: @package_version@
  * @link      http://pear.php.net/package/PHP_CodeSniffer
@@ -70,6 +70,13 @@ class PHP_CodeSniffer_Reporting
      * @var array
      */
     private $_reports = array();
+
+    /**
+     * A cache of opened tmp files.
+     *
+     * @var array
+     */
+    private $_tmpFiles = array();
 
 
     /**
@@ -124,7 +131,7 @@ class PHP_CodeSniffer_Reporting
         $errorsShown = false;
 
         foreach ($cliValues['reports'] as $report => $output) {
-            $reportClass = self::factory($report);
+            $reportClass = $this->factory($report);
 
             ob_start();
             $result = $reportClass->generateFileReport($reportData, $cliValues['showSources'], $cliValues['reportWidth']);
@@ -135,19 +142,22 @@ class PHP_CodeSniffer_Reporting
             $generatedReport = ob_get_contents();
             ob_end_clean();
 
-            if ($generatedReport !== '') {
-                $flags = FILE_APPEND;
-                if (in_array($report, $this->_cachedReports) === false) {
-                    $this->_cachedReports[] = $report;
-                    $flags = null;
+            if ($output === null && $cliValues['reportFile'] !== null) {
+                $output = $cliValues['reportFile'];
+            }
+
+            if ($output === null) {
+                // Using a temp file.
+                if (isset($this->_tmpFiles[$report]) === false) {
+                    $this->_tmpFiles[$report] = tmpfile();
                 }
 
-                if ($output === null) {
-                    if ($cliValues['reportFile'] !== null) {
-                        $output = $cliValues['reportFile'];
-                    } else {
-                        $output = sys_get_temp_dir().'/phpcs-'.$report.'.tmp';
-                    }
+                fwrite($this->_tmpFiles[$report], $generatedReport);
+            } else {
+                $flags = FILE_APPEND;
+                if (isset($this->_cachedReports[$report]) === false) {
+                    $this->_cachedReports[$report] = true;
+                    $flags = null;
                 }
 
                 file_put_contents($output, $generatedReport, $flags);
@@ -179,21 +189,30 @@ class PHP_CodeSniffer_Reporting
         $reportFile='',
         $reportWidth=80
     ) {
-        $reportClass = self::factory($report);
+        $reportClass = $this->factory($report);
 
         if ($reportFile !== null) {
             $filename = $reportFile;
             $toScreen = false;
             ob_start();
-        } else {
-            $filename = sys_get_temp_dir().'/phpcs-'.$report.'.tmp';
-            $toScreen = true;
-        }
 
-        if (file_exists($filename) === true) {
-            $reportCache = file_get_contents($filename);
+            if (file_exists($filename) === true) {
+                $reportCache = file_get_contents($filename);
+            } else {
+                $reportCache = '';
+            }
         } else {
-            $reportCache = '';
+            if (isset($this->_tmpFiles[$report]) === true) {
+                $data        = stream_get_meta_data($this->_tmpFiles[$report]);
+                $filename    = $data['uri'];
+                $reportCache = file_get_contents($filename);
+                fclose($this->_tmpFiles[$report]);
+            } else {
+                $reportCache = '';
+                $filename    = null;
+            }
+
+            $toScreen = true;
         }
 
         $reportClass->generate(
@@ -216,7 +235,7 @@ class PHP_CodeSniffer_Reporting
 
             $generatedReport = trim($generatedReport);
             file_put_contents($reportFile, $generatedReport.PHP_EOL);
-        } else if (file_exists($filename) === true) {
+        } else if ($filename !== null && file_exists($filename) === true) {
             unlink($filename);
         }
 
@@ -252,6 +271,10 @@ class PHP_CodeSniffer_Reporting
 
         // Merge errors and warnings.
         foreach ($phpcsFile->getErrors() as $line => $lineErrors) {
+            if (is_array($lineErrors) === false) {
+                continue;
+            }
+
             foreach ($lineErrors as $column => $colErrors) {
                 $newErrors = array();
                 foreach ($colErrors as $data) {
@@ -261,7 +284,7 @@ class PHP_CodeSniffer_Reporting
                                     'severity' => $data['severity'],
                                     'type'     => 'ERROR',
                                    );
-                }
+                }//end foreach
 
                 $errors[$line][$column] = $newErrors;
             }//end foreach
@@ -270,6 +293,10 @@ class PHP_CodeSniffer_Reporting
         }//end foreach
 
         foreach ($phpcsFile->getWarnings() as $line => $lineWarnings) {
+            if (is_array($lineWarnings) === false) {
+                continue;
+            }
+
             foreach ($lineWarnings as $column => $colWarnings) {
                 $newWarnings = array();
                 foreach ($colWarnings as $data) {
@@ -279,7 +306,7 @@ class PHP_CodeSniffer_Reporting
                                       'severity' => $data['severity'],
                                       'type'     => 'WARNING',
                                      );
-                }
+                }//end foreach
 
                 if (isset($errors[$line]) === false) {
                     $errors[$line] = array();
